soc: silabs: siwx91x: transform nwp soc files into a driver
The goal of this patch is to switch from the nwp.c and nwp.h soc files
to the new nwp driver. During this transition, we also renamed
CONFIG_WISECONNECT_NETWORK_STACK to CONFIG_SILABS_SIWX91X_NWP which are
a better naming to let the user knows that the network coprocessor files
will be added to the compilation.

The switch from a soc file to a driver device introduce a notion of nwp
device that allows us to check for good initialization and ressources
allocation.

Before this patch, it is not possible to know if the nwp have booted
successfully or not. We can now check if the device driver is ready
or not before trying to do operation related to the nwp.

(cherry picked from commit ba1d267c620151152e05bb72439e5b4ce4f9c54c)

Original-Signed-off-by: Martin Hoff <martin.hoff@silabs.com>
GitOrigin-RevId: ba1d267c620151152e05bb72439e5b4ce4f9c54c
Cr-Build-Id: 8700448230302600001
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8700448230302600001
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I7d6f8f58bdde46c66ccc5eefed6cb7f070a0bff8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7065151
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>