driver: clock_control: mspm0: Allow setting SYSOSC clk and restrict MDIV
According to the TI MSPM0 reference manual MDIV must not be set if
SYSOSC is not configured to 4MHz. To make use of this setting, it is
necessary to enable configuring the SYSOSC clock of either 32MHz
(default) or 4MHz with the clock-frequency parameter of the
sysosc node.
If the SYSOSC is configured to run at 4MHz, then the MCLK divider
setting is applied.

(cherry picked from commit 0745f05380d697d86147eddbf3bc2ef7775459a3)

Original-Signed-off-by: Philipp Miedl <phmi@bang-olufsen.dk>
GitOrigin-RevId: 0745f05380d697d86147eddbf3bc2ef7775459a3
Cr-Build-Id: 8698115358104767297
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8698115358104767297
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I46e5098e06d0d8cfe3d1fe5553e93bae683883bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7158039
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>