include: dt-bindings: clock: stm32: rework STM32_DT_CLOCK_SELECT details
Reworks the way the STM32_DT_CLOCK_SELECT builds its elements. Instead of
taking a mask, it takes the MSB of the field. From the MSB and LSB, we
calculate the width of the field, and this width is then stored (instead
of the mask). This allows to gain space for higher values for the fields.

This larger space is necessary to add the selection of the ADC prescaler
on STM32N6 because it is an 8-bit long field.

The allowed width is from 1 to 8 (and internally stored as 0-7 to fit on
3 bits).

STM32_DT_CLKSEL_MASK_GET keeps the same name, since we still need the mask,
and returns the bitmask from the width with the BIT_MASK macro. Other
STM32_DT_CLKSEL_MASK_* macros are renamed with WIDTH.

All call to STM32_DT_CLOCK_SELECT are updated to reflect the change and use
a width instead of a mask.

This also fixes a few issues like STM32H7 MCO macros, and adds MCO_PRE for
STM32N6.

(cherry picked from commit d891aad2ed0f846a9220ef86d4123e4884e5f5c7)

Original-Signed-off-by: Guillaume Gautier <guillaume.gautier-ext@st.com>
GitOrigin-RevId: d891aad2ed0f846a9220ef86d4123e4884e5f5c7
Cr-Build-Id: 8700108490016058097
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8700108490016058097
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I60c45cf1d02017b22bbe0a504d349b53dab84a43
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7083487
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Dawid Niedźwiecki <dawidn@google.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: Jeremy Bettis <jbettis@chromium.org>