drivers: modem: sim7080: Fix invalid call to strtok_r
The second parameter to strtok_r cannot be null.
It is the separator strtok is meant to search for, and it is
not kept anywhere in the saved state (the saved state
is just a pointer to the next piece of the original string).

(cherry picked from commit d6420bb1c7b51365ef1cb3e6873269a7d7820b67)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: d6420bb1c7b51365ef1cb3e6873269a7d7820b67
Cr-Build-Id: 8700833267110927073
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8700833267110927073
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ied6f3737381f75fa253828c5fe5e803001193c6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7050387
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>