drivers: eth_nxp_enet: Fix power mode control
The driver should not take the whole interface down and re-initialize on
every low power entry and exit. This is a lot of latency for no real gain
as far as I can tell. We can just do as the reference manual actually says
which is to set the sleep enable bit to put the module to sleep while still
being able to detect magic packets for wake on LAN.

Also, the only platform that this power action was "enabled" for was
kinetis, but that platform does not have any power management enabled in
Zephyr.  Which means this code was never getting called even with all the
PM configs on.  So basically this code is dead code. But it could be useful
for other platform, such as RT, so there's no reason not to remove the
dependency on kinetis and let it be used for any of the platform as long as
PM_DEVICE is enabled (hence the imply).

(cherry picked from commit bc9869851b5edca1c351cccf8f2aba1a59d747f3)

Original-Signed-off-by: Declan Snyder <declan.snyder@nxp.com>
GitOrigin-RevId: bc9869851b5edca1c351cccf8f2aba1a59d747f3
Cr-Build-Id: 8700380283100427425
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8700380283100427425
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: If74c1f615191c6382f64945cb0d46558c4649a54
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7069691
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>