net: tc-mapping: Fix SKIP_FOR_HIGH_PRIO
Adjust the way the SKIP option worked. Before this patch, a constant
priority offset was considered "high priority" this had the effect, that
the threads assigned to work on the priority were effectively usesless.
To fix it, consider this immediate handling as a pseudo-queue and compute
the tc-thead-mapping based on the effective count (+1 if skipping is
enabled). This makes it so that all threads are usefull and the
high-priority skip-path is considered as a pseudo tc-thread.

(cherry picked from commit 80d00b5ec104be2290b3b6db0869a88fb2c14829)

Original-Signed-off-by: Cla Mattia Galliard <clamattia@gmail.com>
GitOrigin-RevId: 80d00b5ec104be2290b3b6db0869a88fb2c14829
Cr-Build-Id: 8700289681236179201
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8700289681236179201
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I7f6e32f394fd668140abf1cf9c72220eec361037
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7076378
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>