modules: tf-m: Remove some download Kconfigs
Removes two Kconfig which seemed to indicate downloading of a
project would happen automatically, which does not abide by how
to get additional module code in Zephyr. Due to TF-M always setting
these to "DOWNLOAD" in the repo, they are set even if the modules
do not exist so that they do not download e.g. in CI. Unfortunately
it seems that the qcbor one cannot be removed at this time due to
being needed in some applications and is not apache licensed,
though instructions should be provided to users instead describing
how to add it to a module manifest instead, in a later task

(cherry picked from commit 298a35b7140e36c9404feba2c13873d5cffa933f)

Original-Signed-off-by: Jamie McCrae <jamie.mccrae@nordicsemi.no>
GitOrigin-RevId: 298a35b7140e36c9404feba2c13873d5cffa933f
Cr-Build-Id: 8700108490016058097
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8700108490016058097
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I3515799cd003a43bc6d8a1f4e46b78467f6a05e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7083442
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>