net: tcp: Close all connections when interface goes down
If the network interface goes down, close all TCP connections
that are bound to that interface. The reasoning here is that we
need a way to remove IP address from the interface and it might
not be possible if there is a TCP socket that is bound to that address.
If the interface comes back on, we might get the same IP address in
which case the socket closure was not really needed but it is not
possible to know in advance.

(cherry picked from commit ec3bcd39307abbb9c5df0b2d100706ba27ffe9e4)

Original-Signed-off-by: Jukka Rissanen <jukka.rissanen@nordicsemi.no>
GitOrigin-RevId: ec3bcd39307abbb9c5df0b2d100706ba27ffe9e4
Cr-Build-Id: 8699814051879403841
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8699814051879403841
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ie017d95b6714614663d4233b4610c5f6401624ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7089398
Reviewed-by: Dawid Niedźwiecki <dawidn@google.com>
Reviewed-by: Jon Murphy <jpmurphy@google.com>
Commit-Queue: Dawid Niedźwiecki <dawidn@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>