drivers: pwm: stm32: add device tree configuration for deadtime
When using an stm32 in a bridge circuit with complementary outputs,
the deadtime needs to be configurable to avoid shoot-thru current
on the circuit. So, the HAL has the configuration in the BDTR init
and use that api access to set the configuration.

(cherry picked from commit 76c1d713c1a2a3559355d54acf22723311c8fad4)

Original-Signed-off-by: Charles Hardin <ckhardin@gmail.com>
GitOrigin-RevId: 76c1d713c1a2a3559355d54acf22723311c8fad4
Cr-Build-Id: 8697685022200363873
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8697685022200363873
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ie2b509becae9fb49c98fe2bc12237f7ef99d5e86
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7176865
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>