random: rename CSPRNG_AVAILABLE to ENTROPY_NODE_ENABLED
The initial goal of CONFIG_CSPRNG_AVAILABLE was to declare that there was
an entropy node declared in the DT so that if CONFIG_ENTROPY_GENERATOR
was enabled, then the corresponding driver could be enabled.
This works in most of the cases, but not all. A counter example is the
"entropy_bt_hci" driver which is only enabled if CONFIG_BT_HCI_HOST is also
enabled. This means that in a build where the BT stack is not enabled
that driver won't be enabled even if the DT node is still present.

An example of this are the nrf5340[dk|bsim] boards.

(cherry picked from commit f95e815a1a967b68ae65b71ea7891bd3278cc853)

Original-Signed-off-by: Valerio Setti <vsetti@baylibre.com>
GitOrigin-RevId: f95e815a1a967b68ae65b71ea7891bd3278cc853
Cr-Build-Id: 8698275913240343025
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8698275913240343025
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I1f0fc1abf678003ad2d78b8bd34b9c6bf1fe394c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7154372
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>