net: mqtt_sn: Disable support for GWINFO messages
The MQTT-SN v1.2 [0] specification specified the GwAddr as follows:
The GwAdd field has a variable length and contains the address of a GW. Its
depends on the network over which MQTT-SN operates and is indicated in the
first octet of this field. For example, in a ZigBee network the network
address is 2-octet long.

It specifies neither the possible values for the first octet, nor the
format of the network-specific address that follows. Thus, the
specification is incomplete and this functionality unusable.

I also wasn't able to find any implementation which implements a gwinfo
message where the address length is not zero. This includes

The current implementation in Zephyr simply copies a `struct sockaddr`
into GwAddr. This is a bad idea for many reasons, the most important one
being that the format is not even specified by POSIX [1]. They only say,
which defines must exist, not what their values are. And in fact, these
even differ between Zephyr and Linux.

Thus, I think it's best to remove the implementation to prevent people from
using this, which may even lead to memory safety issues, depending on the
length of CONFIG_MQTT_SN_LIB_MAX_ADDR_SIZE. If we were to receive an
updated specification, all we'd have to do is to convert between `struct
sockaddr` and mqtt-sn addresses both ways.

[0] https://groups.oasis-open.org/higherlogic/ws/public/download/66091/MQTT-SN_spec_v1.2.pdf
[1] https://pubs.opengroup.org/onlinepubs/9799919799/basedefs/sys_socket.h.html

(cherry picked from commit 4759e3912237974f0bdd3d640e4895f97ed88c98)

Original-https: //github.com/eclipse-paho/paho.mqtt-sn.embedded-c .
Original-Signed-off-by: Michael Zimmermann <michael.zimmermann@sevenlab.de>
GitOrigin-RevId: 4759e3912237974f0bdd3d640e4895f97ed88c98
Cr-Build-Id: 8695397448042487425
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8695397448042487425
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I1bda7741c6135a1fa7695f1a9a00342bcdb3c1ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7260808
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>