drivers: udc_dwc2: avoid register read on disabled controller
The VBUS bounces when the USB connector is plugged in. This can lead to
events VBUS removed and Suspended occurring in that order. With
hibernation support enabled, hibernation request, as result of the
suspend interrupt, will be delegated to the driver thread. Once the
driver thread is scheduled to process hibernation request, the
controller may be already disabled and controller/phy clocks be off. On
nRF54LM20 this leads to CPU crash and a hang.

To avoid this happening, cancel the possible hibernation request after
interrupts are disabled. Although the thread could be still waked up
because of USBSUSP interrupt raised and event posted, this is considered
to be harmless, as there would be a second check in the thread whether
the event is still valid or is cleared in between.

(cherry picked from commit 0d515e777fb082ad144625c2fb9d9001157ba7ff)

Original-Signed-off-by: Johann Fischer <johann.fischer@nordicsemi.no>
GitOrigin-RevId: 0d515e777fb082ad144625c2fb9d9001157ba7ff
Cr-Build-Id: 8697050844105733473
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8697050844105733473
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I3ba709fcbce5a82f0052a7be1b5b746aabd24c8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7209002
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>