soc: intel_adsp: SoC specific cached/uncached regions support
This re-implements the CONFIG_XTENSA_RPO_CACHE and its cached
and uncached region support in the SoC layer. This is in
preparation for removing RPO cache in the architecture layer
as this is a SoC feature and not an architecture feature.

(cherry picked from commit a82a09a3148c3ad8825102807071f21f59cd0fd9)

Original-Signed-off-by: Daniel Leung <daniel.leung@intel.com>
GitOrigin-RevId: a82a09a3148c3ad8825102807071f21f59cd0fd9
Cr-Build-Id: 8695963679748276289
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8695963679748276289
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Iaf3e5509241028f6f420a7a4b62dda1354bb2d5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7241848
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Dawid Niedźwiecki <dawidn@google.com>
Reviewed-by: Dawid Niedźwiecki <dawidn@google.com>