drivers: sensor: ti: Add INA232 native support
The INA232 is another device in the INA2XX family and is very similar
to the already implemented INA230 and INA236. The main difference
between the INA232 and the INA236 is that the INA232 does not have a
Device ID register. Because of these similarities, it is implemented
in the ina230.x files in the same way as the IN236. Modifications to
the corresponding tests are included as well.

(cherry picked from commit 8aaa5031f7e481912ba234de9bf96e1f4cee56be)

Original-Signed-off-by: Johannes Meyer <johannes.meyer@intego.de>
GitOrigin-RevId: 8aaa5031f7e481912ba234de9bf96e1f4cee56be
Cr-Build-Id: 8696620508454174593
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8696620508454174593
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I611a5c38ad1b4c846ea063e03cc0f649e327c6c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7214278
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>