tests: spi_loopback: Skip tests properly
There is a couple test cases which for whatever reason does not work on
certain platform drivers, and they are being hardcoded to be skipped in
the test if using those drivers. This decision was made a long time ago,
and I do not have great insight into why, but regardless, the proper way
to skip a ztest case is to call the proper ztest skip API instead of
just returning, because it should be marked as skipped, not passed.
Also, the problem for these couple drivers is clearly the cases where rx
is bigger than tx, so add the same skip for the other new case where rx
is bigger than tx, the same_buf_cmd case.

After analyzing the DSPI driver a bit, I have a suspicion that this is
just a software driver limitation, that could be overcome to function
properly, and not a hardware limitation. But this commit is just to skip
the test properly that is already being skipped, not fix the underlying
issues with these drivers.

(cherry picked from commit b5de9fe252666ebf5fcd7623935f46a6332377d7)

Original-Signed-off-by: Declan Snyder <declan.snyder@nxp.com>
GitOrigin-RevId: b5de9fe252666ebf5fcd7623935f46a6332377d7
Cr-Build-Id: 8698275913240343025
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8698275913240343025
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ia844752b1e2bdc79c29e09182e2deaf362bdcd5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7154416
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>