kernel: Adjust metairq preemption tracking bounds
Adjust the bounds for tracking metairq preemption to include the
case where the number of metairq threads matches the number of
cooperative threads. This is needed as a thread that is schedule
locked through k_sched_lock() is documented to be treated as a
cooperative thread. This implies that if such a thread is preempted
by a metairq thread that execution control must return to that
thread after the metairq thread finishes its work.

(cherry picked from commit 653731d2e11fff87a8646c1297b986c150491433)

Original-Signed-off-by: Peter Mitsis <peter.mitsis@intel.com>
GitOrigin-RevId: 653731d2e11fff87a8646c1297b986c150491433
Cr-Build-Id: 8695850433277688401
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8695850433277688401
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I975ab16b70013cdb9ed6aecc6960a00e5eac10c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7248553
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>