display: stm32-ltdc: clarify pixel-format property of lcd_controller
So far the LTDC driver has not be using the pixel-format property
defined in the lcd_controller bindings. This property describes the
format of the panel connected to the LTDC output. It is not to be
confused with the format of the LTDC layers (aka format of data stored
into the framebuffer). The LTDC accepts various input formats while in
most cases only RGB888 is supported for output. RGB666 is also supported
by enabling Dithering however it is not supported currently by the
driver.
Enforce verification of the pixel-format given to the LTDC device-tree
by checking that it matches with the platform supported formats.

(cherry picked from commit 73ca8c9f76d3de661178ee1b8431a2c7dc3429a8)

Original-Signed-off-by: Alain Volmat <alain.volmat@foss.st.com>
GitOrigin-RevId: 73ca8c9f76d3de661178ee1b8431a2c7dc3429a8
Cr-Build-Id: 8695941030404006097
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8695941030404006097
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I0fc77c6bcc5ca8a72db7ab6e7e417ce703d37c00
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7243004
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Dawid Niedźwiecki <dawidn@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>