docs: bindings-syntax: re-add note about `title` being optional
When originally introduced in [1], `title` was described as optional, but
this information was lost when the documentation was reworked in [2], even
though the underlying behavior in edtlib is unchanged.

Update the documentation to indicate that that `title` is optional. While
at it, fix a little typo in modified paragraph.

[1]: ee17657ad3290465e8255c8c9da7efbd5343f789
[2]: 3a2f839d459ae20597f1bc4af0e78d35a68b5dd4

(cherry picked from commit d770c6ce13aa735a6406c18be99a7206de94995b)

Original-Signed-off-by: Mathieu Choplain <mathieu.choplain-ext@st.com>
GitOrigin-RevId: d770c6ce13aa735a6406c18be99a7206de94995b
Cr-Build-Id: 8696529912005117457
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8696529912005117457
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I375ec399a772defbb036fdbc6bc91462f9066ad0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7219685
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>