Bluetooth: Classic: HFP_AG: Optimize audio connect management
In current implementation, the call will be terminated, or rejected if
the audio connect cannot be established. But from the application
view, if the sco connect cannot be established, the audio paths can be
routed towards the AG. So force routeing the audio towards the HF is
not reasonable and it is not the mandatory.

The changes aim to keep the sequence that creating SCO connect in the
right time. And ignore the result whether the SCO connect has been
created or can be established.

If the audio connect cannot be established, the application can try to
change the audio route paths according to the audio connection
procedure.

(cherry picked from commit 741afc85af2a310a0bc8ac4de15b9c5813f58449)

Original-Signed-off-by: Lyle Zhu <lyle.zhu@nxp.com>
GitOrigin-RevId: 741afc85af2a310a0bc8ac4de15b9c5813f58449
Cr-Build-Id: 8697050844105733473
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8697050844105733473
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I2dc9af630b0c56c2b1e3b090efa32f52054976ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7209021
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>