net: tcp: Wake pending threads if non-blocking connect fails
If a thread was sleeping in poll(), monitoring POLLIN only, it should
still be notified if an asynchronous connection, triggered by
non-blocking connect(), failed. Currently that's not the case.
In result the application would never know whether the connection
was successful or not and would be stuck with a disconnected socket
that would not report anything with poll().

This was not an issue if POLLOUT was monitored as well, because POLLOUT
sets up a connect semaphore that was reset in case of errors. POLLIN
however only monitors the recv fifo, which was not waken up in such
case.

Fix this by canceling any pending waits on recv fifo and recv
condition variable.

(cherry picked from commit 6be292c35efc4063235587d02794cb182d602d0f)

Original-Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>
GitOrigin-RevId: 6be292c35efc4063235587d02794cb182d602d0f
Cr-Build-Id: 8697571776524936913
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8697571776524936913
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Iae02e80d6cfe122f5e47d7c0f29025559b217d54
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7185971
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>