tests: net: socket: tcp: Add another test case for asynchronous connect
Verify that if the application triggers asynchronous connect with
non-blocking sockets, and starts to monitor the socket with poll()
immediately but with POLLIN only set, the poll() is still able to
report an error if the connection fails. As no POLLOUT is monitored, the
application won't know when the connection is done in case of success,
but it should still be notified in case of errors.

To achieve this, modify the existing test case to allow to specify what
events should be monitored by poll(). Additionally monitor the time
spent in poll() - an error should cause poll() to exit immediately,
not after the specified timeout.

(cherry picked from commit 396fcbf4f53cf42c9de4e415110f06016cab7bc1)

Original-Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>
GitOrigin-RevId: 396fcbf4f53cf42c9de4e415110f06016cab7bc1
Cr-Build-Id: 8697571776524936913
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8697571776524936913
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I98cbc115d96b0d7cfdc7cbb377525f31c3462483
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7185972
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>