debug: coredump: stack top limit for current thread
Extend the functionality to limit the number of stack bytes
included in the core dump by allowing the limit to be
different for the current thread and remaining threads.

This is useful because it is more likely that we need more
call frames of the thread that was running when the
exception occurred than of the other threads in order to
analyze the exception.

(cherry picked from commit 6402a58f34ae7501a4d9726f1056a9b24497345e)

Original-Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>
GitOrigin-RevId: 6402a58f34ae7501a4d9726f1056a9b24497345e
Cr-Build-Id: 8696484613249217009
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8696484613249217009
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I1fc06cd95f088a5168970acbfc25ff9ebfcb30c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7225011
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>