Bluetooth: BAP: Fix bad check for ASE state for CIS connect
A CIS may be connected in either the QoS Configured state
or the enabling state. The QoS Configured state is the
earliest state it is allowed, due to it being the first
state where the CIS_ID and ASE_ID are paired.
The enabling state is the "last" state it is allowed,
as if the ASE is in the streaming, disabling or releasing
state we should not allow/expect a CIS connection to happen.

(cherry picked from commit 4c54eef0d0f00c32d075a7498ad33d41b2764cb8)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 4c54eef0d0f00c32d075a7498ad33d41b2764cb8
Cr-Build-Id: 8697299984559700737
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8697299984559700737
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ieb494e0b6c0bbb28819522694bc5c9c21f59b72a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7199740
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>