net: l2: ieee802154: add support for MLME coordinator address
There are two ways in the current implementation to join a network,
either with an association request or a via MLME-SET(macShortAddress) to
rejoin a previously associated network.

The coordinator address needs to be valid in the rejoin case. It can be
lost, for example if the device power cycles. This commit fixes that gap
with new MLME GET/SET for:
- macCoordShortAddress
- macCoordExtendedAddress

(cherry picked from commit f186218e4285a3730a91de6ac96f96128270a2c8)

Original-Signed-off-by: Simon Piriou <spiriou31@gmail.com>
GitOrigin-RevId: f186218e4285a3730a91de6ac96f96128270a2c8
Cr-Build-Id: 8696665805680353105
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8696665805680353105
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ib89cb0cd4a1cf6a311a868f95f05f8dcb84a3129
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7213951
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>