scripts: west: nrfutil: Use build dir to store generated json
Until now the code was using the path to the .hex file to select the
directory where the generated JSON file required for nrfutil would be
generated. But this has a problem: if the .hex file in a sysbuild
project is located in the tree as a precompiled binary, the runner would
place a file in there and make the tree dirty. Instead use the build_dir
folder which always corresponds (in both sysbuild and regular builds)
to the current target build directory.

(cherry picked from commit d6f5aa925832800204d3e0fe0015a924de5323f6)

Original-Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>
GitOrigin-RevId: d6f5aa925832800204d3e0fe0015a924de5323f6
Cr-Build-Id: 8697571776524936913
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8697571776524936913
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I6a41fe8bfa6bf6526c37aad04380ea365d3e5618
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7185964
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>