samples: st: power_mgmt: wkup_pins: correct overlay for Nucleo-F103RB
The overlay indicated the pin was active-low with a pull-up resistor, but
the hardware forces a pull-DOWN which makes the pin active-HIGH instead.
Also add a comment indicating that the pin is not wired to a button and
a short (to Vdd!) must be applied via a jumper wire.

(cherry picked from commit 3db76f037391b5f4383e67ac876ef921b910b23d)

Original-Signed-off-by: Mathieu Choplain <mathieu.choplain-ext@st.com>
GitOrigin-RevId: 3db76f037391b5f4383e67ac876ef921b910b23d
Cr-Build-Id: 8697209811984320625
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8697209811984320625
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I7f26a100ce2a7c05cdf66e1b846e2ae19ee6da9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7204181
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>