MCUmgr: OS: fix set datetime millisecond handling
According to the docs the millis were in the format `.SSSSSS`. In
reality though, it only accepted exactly `.SSS`, not `.SS` or `.SSSS`
and specifically also not `.SSSSSS`, contrary to the docs. Further, it did
not fail with an error message but simply produced the wrong value.

With this change it accepts everything from `.` to `.SSSSSS` and
produces the correct result. This is compatible with the previous
behavior, with the documentation and with everything in between.

(cherry picked from commit ffb046b7971a928bb3a7f32445c35c6c90b476c9)

Original-Signed-off-by: Martin Stumpf <finomnis@gmail.com>
GitOrigin-RevId: ffb046b7971a928bb3a7f32445c35c6c90b476c9
Cr-Build-Id: 8695759836740870721
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8695759836740870721
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I43415e5e71642d07b0583e5a401045680a10c1a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7253676
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>