tests: posix: headers: remove testsuite
Remove the tests/posix/headers testsuite.

A large part of this testsuite was originally feeling-out what types,
functions, and other declarations were expected in POSIX headers when
CONFIG_POSIX_API was not selected.

Really, libraries and applications should generally not expect to be able
to use POSIX headers in particularly meaningful ways when POSIX APIs are
not enabled.

The portions of the testsuite that check for the existence of declarations
and symbols when POSIX APIs are configured also effectively duplicates
the implicit tests of simply compiling code and testsuites.

Also, given that CONFIG_POSIX_API is effectively deprecated, it does not
make a lot of sense to maintain the testsuite.

(cherry picked from commit 94e07d235b8bb32c169ee735b81b93caefacc285)

Original-Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>
GitOrigin-RevId: 94e07d235b8bb32c169ee735b81b93caefacc285
Cr-Build-Id: 8697209811984320625
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8697209811984320625
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I3ad05b7b9557e6f50d0ebc18afb9edd28ae06a40
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7204182
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>