soc: nxp: imxrt: Initialize FlexRAM before first stack use
If the stack is in ITCM, DTCM, or FlexRAM OCRAM, flexram_dt_partition()
may change its contents. The comment on flexram_dt_partition()
acknowledges that, stating that it's inlined because it "cannot use
[the] stack". But we currently call SystemInit(), which is not inlined
and does use the stack, prior to flexram_dt_partition()! Fix that issue
by reordering the calls.

It seems to me that flexram_dt_partition() would be safer as a
soc_early_reset_hook implemented in assembly, but this fix does work for
the moment. Tested on an i.MX RT1061, with the stack in FlexRAM OCRAM
and entering Zephyr with all FlexRAM allocated to ITCM and DTCM.

(cherry picked from commit 799ac970d48023a806afb8e43fea22ccaf185e9a)

Original-Signed-off-by: Thomas Hebb <tommyhebb@gmail.com>
GitOrigin-RevId: 799ac970d48023a806afb8e43fea22ccaf185e9a
Cr-Build-Id: 8696212821870935201
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8696212821870935201
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I0421154bc41636cb07789383e296d3411df9b9f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7233601
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Dawid Niedźwiecki <dawidn@google.com>