Bluetooth: Classic: A2DP: Fix out of bounds issue
This is a negative detection. The conn index should be never out of
bounds of the array `connection`. Because the size of the array is
`CONFIG_BT_MAX_CONN`.

To improve the readability, add a `__ASSERT` checking to check if the
conn index is out of bounds.

CID 487769

(cherry picked from commit 05dc860f119fe50de731f7fef6474cd89f20bbdb)

Original-Signed-off-by: Lyle Zhu <lyle.zhu@nxp.com>
GitOrigin-RevId: 05dc860f119fe50de731f7fef6474cd89f20bbdb
Cr-Build-Id: 8696665805680353105
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8696665805680353105
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I950d59aaeb8ebfacd1334ec9c2b19c78ca10fac2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7213307
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>