tests: Bluetooth: Tester: Enable CAP commander and handover
Enable the CAP Commander role for the BT tester. This will
help pass some TMAP and GMAP tests that require the CAP
Commander to be enabled.

The actual implementation of the tests related to the
CAP commander is not implemented.

TMAP also requires that handover is supported if we
support both unicast and broadcast, so enable the config
for now. Support for the handover tests will follow in
a separate commit.

(cherry picked from commit ca085cee3c00b64080c944cf484cf999adffe8a9)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: ca085cee3c00b64080c944cf484cf999adffe8a9
Cr-Build-Id: 8696054276481881393
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8696054276481881393
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I7e2d1a91f5ca2ca30972190655c451074e8efe94
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7237444
Reviewed-by: Dawid Niedźwiecki <dawidn@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Dawid Niedźwiecki <dawidn@google.com>