cmake: FindHostTools: Do not override cross-compile for native builds
This check is meant to be easier on users so they do not need to bother
setting ZEPHYR_TOOLCHAIN_VARIANT to "host" when building for
native_simulator based targets, so we default in that case to it.
But when somebody has actively set it to "cross-compile" we should not
override it.

(cherry picked from commit 654535bea38d9b115d5879dbc5d5417240d269d7)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: 654535bea38d9b115d5879dbc5d5417240d269d7
Cr-Build-Id: 8696326068362242273
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8696326068362242273
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I2497101ea730e0956b96d161e774ba97140c7569
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7231822
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>