drivers/flash/flash_simulator: Add write and erase callbacks
Add the functionality that allows to register write and erase callbacks.
The callbacks can modify the behaviour of this operations.
It are meant to be used during testing of the situation when the memory
starts to generate errors or the data cannot be trusted anymore.
It expands the testing possibility, for a situation when the module
is tested that checks the data integrity and and we cannot just jump in
between its internal write and read operations to inject an error.

(cherry picked from commit 63c3c34f74227847215467f50b80f25af6248c4b)

Original-Signed-off-by: Radosław Koppel <r.koppel@k-el.com>
GitOrigin-RevId: 63c3c34f74227847215467f50b80f25af6248c4b
Cr-Build-Id: 8695035061077862097
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8695035061077862097
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I38021fdcf54dde806c3bfd08554434e5dfcd0297
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7280409
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Commit-Queue: Ting Shen <phoenixshen@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>