drivers: modem: Include empty string arguments in argument parsing
When parsing AT command responses, empty string arguments were being
ignored if they occur it the end of the response.
This is not only iconsistent behavior, but also causes issues when
parsing responses where we need to know the number of arguments.
For example, the +CEREG response and unsolicited command response which
we need to know if it is response or unsolicited command response to
correctly parse the it, as the first argument will be different.
Here we want to know that response has 8 arguments, not 7, or we will
confuse it with the notification.

(cherry picked from commit 79b90253151e0b3b79cf8c6dd2e1b11b6dd1aac5)

Original-Example:
Original-Response: +CEREG: 5,4,\"0A0B\",\"FFFFFFFF\",9,0,0,,
Original-Notification: +CEREG: 4,\"0A0B\",\"FFFFFFFF\",9,0,0,,
Original-Signed-off-by: Joakim Andersson <joerchan@gmail.com>
GitOrigin-RevId: 79b90253151e0b3b79cf8c6dd2e1b11b6dd1aac5
Cr-Build-Id: 8695941030404006097
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8695941030404006097
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I3da9c465958b354f8041524da30684d1b03676e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7243023
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Dawid Niedźwiecki <dawidn@google.com>