scripts: build: check_init_prio: take deferred-init into account
A device with zephyr,deferred-init is initialized by the application, so
we cannot check if it is initialized in "the correct order". On the other
hand, a non-deferred device must not depend on a deferred-init device: this
is guaranteed to be initialized out of order.

Also adjust existing test cases and add a new test case to validate this
behavior.

(cherry picked from commit cb27d92c0c139727857122add232cfd7c8791ce3)

Original-Signed-off-by: Mathieu Choplain <mathieu.choplain-ext@st.com>
GitOrigin-RevId: cb27d92c0c139727857122add232cfd7c8791ce3
Cr-Build-Id: 8695827784712109121
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8695827784712109121
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I69db35051bc640815ca624cf36371e5ba8cc3640
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7249325
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>