llext: Keep constant data in flash.
This patch adds a new Kconfig option that allows read-only data sections
without relocations to be mapped directly (remain in flash) instead of
being copied to RAM during extension loading. This helps extensions that
have large constant data, such as certificates, lookup tables etc.. to
save saving significant RAM, especially with MPU is enabled.

This feature requires the extension to place rodata sections without
relocations in LLEXT_RODATA_NO_RELOC section.

Note that we chose to disable this optimization when MMU or USERSPACE is
enabled because the NO_RELOC section may not be aligned to the MPU/MMU
requirements on some architectures. If we leave it in place, llext will
attempt to create an MPU region for it using a misaligned address, and
if we allow it to be copied to RAM, it defeats the purpose of this feature.

(cherry picked from commit dca4136f5e07dafefb2f116d529bebb502e678cd)

Original-Signed-off-by: Ibrahim Abdalkader <i.abdalkader@gmail.com>
GitOrigin-RevId: dca4136f5e07dafefb2f116d529bebb502e678cd
Cr-Build-Id: 8692588942390808017
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8692588942390808017
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I8435515a58df2f7b6bc7df6f065fadbe6fd7528a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7487559
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>