drivers: disk: loopback: Remove unnecessary include
This file does not seem to need unistd.h, and that header is a
POSIX extension to the C library which is not avaiable in general.
So let's not include it.

Without this change this file fails to build with minimal libc or
other C libraries which do not have this extra header.

(cherry picked from commit 9f0ce75b411af1d03663b4302d16de07ef223c52)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: 9f0ce75b411af1d03663b4302d16de07ef223c52
Cr-Build-Id: 8695827784712109121
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8695827784712109121
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I33e831a388ad4e08933c0dee33be09f6aaceb6f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7249327
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>